Structurize benchmarks
