+++ benchmarks/GeneratedHydratorBenchmark/Data/InheritanceDeepClass.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace GeneratedHydratorBenchmark;
+namespace GeneratedHydratorBenchmark\Data;
 
 /**
  * Class that contains mixed deeply inherited properties
